refactor url download functions into internet subtable
